+++ packages/terra-form-textarea/src/Textarea.jsx
@@ -218,6 +218,7 @@ class Textarea extends React.Component {
 
 Textarea.propTypes = propTypes;
 Textarea.defaultProps = defaultProps;
+Textarea.isTextarea = true;
 
 export default Textarea;
 export { TextareaSize };