+++ packages/terra-form-input/CHANGELOG.md
@@ -3,6 +3,8 @@ ChangeLog
 
 Unreleased
 ----------
+### Changed
+* Added `isInput` type on Input component to help with component identity
 
 2.12.0 - (May 1, 2019)
 ------------------